fix event manager
